单位趋势列表

+++ ml-firecontrol/ml-firecontrol-dto/src/main/java/com/mailian/firecontrol/dto/web/response/BgUnitTrendResp.java
@@ -3,14 +3,12 @@ package com.mailian.firecontrol.dto.web.response;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
-import java.math.BigDecimal;
-
 @ApiModel(description = "大屏单位走势")
 public class BgUnitTrendResp {
     @ApiModelProperty(value = "日期")
     private String date;
     @ApiModelProperty(value = "数值")
-    private BigDecimal num;
+    private float num;
 
     public String getDate() {
         return date;
@@ -20,11 +18,11 @@ public class BgUnitTrendResp {
         this.date = date;
     }
 
-    public BigDecimal getNum() {
+    public float getNum() {
         return num;
     }
 
-    public void setNum(BigDecimal num) {
+    public void setNum(float num) {
         this.num = num;
     }
 }
